// src/app/models/cep.ts
export interface Cep {
    
        cep: string;
        logradouro: string;
        complemento: string;
        bairro: string;
        idCidade: number;
        localidade: string;
        idUf: number;
        uf: string;
        nomeUf: string;
        ibge: string;
        gia: string;
        ddd: string;
        siafi:string;
}